# license-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DB Migrator — License Server API",
    "version": "1.1.0",
    "description": "Central license + billing server。提供帳號 / session / 付款 / 2FA / API tokens / offline-license revocation。\n\n供 DB Migrator client（node-express / dotnet8）連線驗證 license，也供 portal UI 自助管理。",
    "contact": { "url": "https://github.com/wei820528/db-migrator" },
    "license": { "name": "PolyForm Free Trial 1.0.0" }
  },
  "servers": [
    { "url": "http://localhost:4000", "description": "Local dev" },
    { "url": "https://license.example.com", "description": "Production placeholder" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "兩種 token 都接受：\n- 一般 session token（從 /api/auth/login 拿，hex string）\n- API token，prefix `dbmt_`（從 /api/user/tokens 拿，給 cron / CI 用）"
      },
      "adminCookie": {
        "type": "apiKey", "in": "cookie", "name": "admin_session",
        "description": "Admin UI cookie，從 /api/admin/login 拿"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
      "User": {
        "type": "object",
        "properties": {
          "email":        { "type": "string", "format": "email" },
          "plan":         { "type": "string", "enum": ["trial", "basic", "team", "enterprise"] },
          "status":       { "type": "string", "enum": ["trial", "licensed", "free", "expired", "kicked", "offline"] },
          "daysLeft":     { "type": "integer", "nullable": true },
          "features":     { "type": "object", "additionalProperties": true },
          "maxDevices":   { "type": "integer" },
          "totpEnabled":  { "type": "boolean" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": { "description": "未登入或 token 無效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "BadRequest":   { "description": "請求格式錯誤",       "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    }
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"], "summary": "建立 trial 帳號",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["email", "password"],
          "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string", "minLength": 8 } }
        } } } },
        "responses": {
          "200": { "description": "已寄出驗證信（或 dev mode 直接 verified）" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "description": "email 已註冊" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"], "summary": "登入；若 2FA 啟用會回 challenge",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["email", "password"],
          "properties": { "email": { "type": "string" }, "password": { "type": "string" } }
        } } } },
        "responses": {
          "200": {
            "description": "登入成功（或 needs2fa）",
            "content": { "application/json": { "schema": {
              "oneOf": [
                { "type": "object", "properties": { "token": { "type": "string" }, "user": { "$ref": "#/components/schemas/User" } } },
                { "type": "object", "properties": { "needs2fa": { "type": "boolean" }, "challengeId": { "type": "string" } } }
              ]
            } } }
          },
          "401": { "description": "invalid credentials" },
          "403": { "description": "email 尚未驗證 / IP 不在白名單 / plan 過期" }
        }
      }
    },
    "/api/auth/heartbeat": {
      "post": {
        "tags": ["Auth"], "summary": "Client app 心跳；IP 變了會被踢",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "session 有效" },
          "401": { "description": "kicked（IP 變了 / 被 admin 踢 / token 失效）" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"], "summary": "目前 session 詳細",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"], "summary": "登出（刪 server-side session）",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/auth/verify-email": {
      "get": {
        "tags": ["Auth"], "summary": "Email 驗證 link（從信中點進來）",
        "parameters": [{ "name": "token", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "驗證成功 HTML" }, "404": { "description": "token 無效或已用過" } }
      }
    },
...
    "/api/billing/webhook": {
      "post": {
        "tags": ["Billing"], "summary": "Stripe webhook（raw body，已驗 signature）",
        "responses": { "200": { "description": "OK" }, "400": { "description": "signature 驗證失敗" } }
      }
    },
...
    "/api/admin/webhooks": {
      "get": {
        "tags": ["AdminWebhooks"],
        "summary": "列 admin webhooks + KNOWN_EVENTS",
        "security": [{ "adminCookie": [] }],
        "responses": { "200": { "description": "{ webhooks: [{ id, name, url, events, active, last_event, last_status, ... }], knownEvents: [...] }" } }
      },
      "post": {
        "tags": ["AdminWebhooks"],
        "summary": "建立 webhook — secret 只在 response 回傳一次",
        "description": "HMAC-SHA256 用 secret 簽 body；header `X-LicenseServer-Signature: sha256=<hex>`。Subscriber 端驗章方式同 node-express webhooks (v1.2.0 之前範例)。",
        "security": [{ "adminCookie": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["name", "url", "events"],
          "properties": {
            "name":   { "type": "string", "maxLength": 64 },
            "url":    { "type": "string", "format": "uri", "description": "http:// 或 https://" },
            "events": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["*", "ping", "user.kicked", "user.registered", "license.expired",
                         "payment.succeeded", "payment.failed", "subscription.canceled"]
              },
              "minItems": 1
            }
          }
        } } } },
        "responses": {
          "200": { "description": "{ id, name, url, events, secret (lswhsec_…, 僅此一次), secret_prefix }" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/admin/webhooks/{id}": {
      "patch": {
        "tags": ["AdminWebhooks"], "summary": "更新（name / url / events / active）",
        "security": [{ "adminCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "tags": ["AdminWebhooks"], "summary": "刪除 webhook",
        "security": [{ "adminCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "{ ok: true }" }, "404": { "description": "not found" } }
      }
    },
    "/api/admin/webhooks/{id}/test": {
      "post": {
        "tags": ["AdminWebhooks"], "summary": "Test-ping — 立刻送一個 `ping` 事件，驗 receiver 收得到 + 驗章 OK",
        "security": [{ "adminCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "{ ok, status, attempt } 或 { ok:false, attempts, error }" },
          "400": { "description": "webhook not found" }
        }
      }
    }
  }
}